Fix bug where parents list not cleared
This is a bit of an edge case as it's unlikely that people are likely to
have deeply-nested structures where some bits are mergeable and some
aren't, but it's worthwhile.

Previously the added test case would fail because the `parents` array
wouldn't have been cleared ([:a, :b, :e] vs [:a, :e]).

Now we're combining the parents and keys in the call without affecting
the parents object and the behvaiovur appears corret.